feat: use watch flag on tailwind build
