+++ pages/content-ui/package.json
@@ -13,8 +13,7 @@
     "clean:node_modules": "pnpx rimraf node_modules",
     "clean:turbo": "rimraf .turbo",
     "clean": "pnpm clean:bundle && pnpm clean:node_modules && pnpm clean:turbo",
-    "ready": "tsx build-tw.mts",
-    "build": "pnpm ready && tsx build.mts",
+    "build": "tsx build.mts",
     "dev": "tsx build.mts",
     "lint": "eslint .",
     "lint:fix": "pnpm lint --fix",
